[feat] 회원 접근 페이지에서 로그아웃 시 비정상적인 접근으로 표시되는 오류 해결
