+++ dutchiepay/src/app/(routes)/mart/write/[id]/page.jsx
@@ -5,6 +5,7 @@ import { useEffect, useState } from 'react';
 import { ALL_COMMUNITY_CATEGORIES } from '@/app/_util/constants';
 import LocationModal from '@/app/_components/_community/_local/LocationModal';
 import MartPostForm from '@/app/_components/_community/_local/MartPostForm';
+import ProtectedRoute from '@/app/_components/ProtectedRoute';
 import axios from 'axios';
 import { getStringDateToTimeStamp } from '@/app/_util/getFormatDate';
 import getTextLength from '@/app/_util/getTextLength';
@@ -97,36 +98,37 @@ export default function MartModify() {
 
       router.push(`/mart/${id}`);
     } catch (error) {
-      console.log(error);
       alert('오류가 발생했습니다. 다시 시도해주세요.');
     }
   };
 
   return (
-    <section className="min-h-[750px] w-[1020px] mb-[100px] mt-[40px] mx-[60px] relative">
-      <form onSubmit={handleSubmit(onSubmit)}>
-        <MartPostForm
-          setIsModalOpen={setIsModalOpen}
-          register={register}
-          setValue={setValue}
-          watch={watch}
-          editorContent={editorContent}
-          setEditorContent={setEditorContent}
-          thumbnail={thumbnail}
-          images={images}
-          setImages={setImages}
-          setThumbnail={setThumbnail}
-          isUpdate={true}
-        />
-      </form>
-      {isModalOpen && (
-        <LocationModal
-          setIsModalOpen={setIsModalOpen}
-          register={register}
-          setValue={setValue}
-          watch={watch}
-        />
-      )}
-    </section>
+    <ProtectedRoute>
+      <section className="min-h-[750px] w-[1020px] mb-[100px] mt-[40px] mx-[60px] relative">
+        <form onSubmit={handleSubmit(onSubmit)}>
+          <MartPostForm
+            setIsModalOpen={setIsModalOpen}
+            register={register}
+            setValue={setValue}
+            watch={watch}
+            editorContent={editorContent}
+            setEditorContent={setEditorContent}
+            thumbnail={thumbnail}
+            images={images}
+            setImages={setImages}
+            setThumbnail={setThumbnail}
+            isUpdate={true}
+          />
+        </form>
+        {isModalOpen && (
+          <LocationModal
+            setIsModalOpen={setIsModalOpen}
+            register={register}
+            setValue={setValue}
+            watch={watch}
+          />
+        )}
+      </section>
+    </ProtectedRoute>
   );
 }
